Refactor ThreadPool and SynchronizationContext to prepare for non-WinRT port
